Add default copyright for use with CopyrightWizard plugin. No code mods.
The newly added project copyright is not quite identical to the the
copyright text currently in use but it's as close as I could get it,
given the constraints of CopyrightWizard.

The Files template has also been updated so that any new
Java files automatically get the same copyright notice.

The download site for CopyrightWizard is http://www.wdev91.com/update/